Move idents and body-related parsers into their respective module
